+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 - Add `discovery.Connector.IsFatal`
 - Add `discovery.Inspector.IsFatal`
+- Add `NameRoles.RangeByRole()`
+- Add `TypeRoles.RangeByRole()`
 
 ### Removed